refactor(risk): 重构风险规则系统，迁移A股相关规则到markets模块
重构风险规则系统架构，将A股特定规则迁移至markets/ashare模块。主要变更包括：
1. 移除risk模块中原有的通用规则实现，改为依赖nautilus-rules crate
2. 在markets/ashare中实现A股特有的风险规则和提供者接口
3. 改进市场数据提供者设计，统一规则数据获取方式
4. 优化价格限制计算精度，使用i128避免溢出
5. 添加命令规则链支持，分离撤单会话检查逻辑

+++ crates/markets/ashare/src/market/price_limits.rs
@@ -192,11 +192,11 @@ pub fn compute_price_limits(
         }
 
         // 计算涨停价：向上取整到tick
-        let limit_up_raw = ((prev_close_raw as f64 * (1.0 + limit_config.limit_up_pct) / tick_raw as f64).ceil() as i64) * tick_raw;
+        let limit_up_raw = ((prev_close_raw as f64 * (1.0 + limit_config.limit_up_pct) / tick_raw as f64).ceil() as i128) * tick_raw as i128;
         let limit_up = Price::from_raw(limit_up_raw as PriceRaw, prev_close.precision);
 
         // 计算跌停价：向下取整到tick
-        let limit_down_raw = ((prev_close_raw as f64 * (1.0 - limit_config.limit_down_pct) / tick_raw as f64).floor() as i64) * tick_raw;
+        let limit_down_raw = ((prev_close_raw as f64 * (1.0 - limit_config.limit_down_pct) / tick_raw as f64).floor() as i128) * tick_raw as i128;
         let limit_down = Price::from_raw(limit_down_raw as PriceRaw, prev_close.precision);
 
         PriceLimits::new(Some(limit_up), Some(limit_down))
